#1910: Kecuali admin, sekarang pengguna hanya dapat mengubah artikel yang dibuat oleh pengguna itu sendiri.

+++ donjo-app/controllers/Web.php
@@ -86,6 +86,9 @@ class Web extends CI_Controller {
 
 	public function form($cat = 1, $p = 1, $o = 0, $id = '')
 	{
+		if (!$this->web_artikel_model->boleh_ubah($id, $_SESSION['user']))
+			redirect("web/index/$cat/$p/$o");;
+
 		$data['p'] = $p;
 		$data['o'] = $o;
 		$data['cat'] = $cat;
@@ -139,12 +142,18 @@ class Web extends CI_Controller {
 	}
 
 	public function update($cat = 0, $id = '', $p = 1, $o = 0){
+		if (!$this->web_artikel_model->boleh_ubah($id, $_SESSION['user']))
+			redirect("web/index/$cat/$p/$o");;
+
 		$this->web_artikel_model->update($cat, $id);
 		redirect("web/index/$cat/$p/$o");
 	}
 
 	public function delete($cat = 1, $p = 1, $o = 0, $id = '')
 	{
+		if (!$this->web_artikel_model->boleh_ubah($id, $_SESSION['user']))
+			redirect("web/index/$cat/$p/$o");;
+
 		$_SESSION['success'] = 1;
 		$outp = $this->web_artikel_model->delete($id);
 		if (!$outp) $_SESSION['success'] = -1;
@@ -165,6 +174,9 @@ class Web extends CI_Controller {
 
 	public function ubah_kategori_form($id = 0)
 	{
+		if (!$this->web_artikel_model->boleh_ubah($id, $_SESSION['user']))
+			redirect("web/index");;
+
 		$data['list_kategori'] = $this->web_kategori_model->list_kategori("kategori");
 		$data['form_action'] = site_url("web/update_kategori/$id");
 		$data['kategori_sekarang'] = $this->web_artikel_model->get_kategori_artikel($id);
@@ -180,24 +192,36 @@ class Web extends CI_Controller {
 
 	public function artikel_lock($cat = 1, $id = 0)
 	{
+		if (!$this->web_artikel_model->boleh_ubah($id, $_SESSION['user']))
+			redirect("web/index/$cat");;
+
 		$this->web_artikel_model->artikel_lock($id, 1);
 		redirect("web/index/$cat");
 	}
 
 	public function artikel_unlock($cat = 1, $id = 0)
 	{
+		if (!$this->web_artikel_model->boleh_ubah($id, $_SESSION['user']))
+			redirect("web/index/$cat");;
+
 		$this->web_artikel_model->artikel_lock($id, 2);
 		redirect("web/index/$cat");
 	}
 
 	public function komentar_lock($cat = 1, $id = 0)
 	{
+		if (!$this->web_artikel_model->boleh_ubah($id, $_SESSION['user']))
+			redirect("web/index/$cat");;
+
 		$this->web_artikel_model->komentar_lock($id, 0);
 		redirect("web/index/$cat");
 	}
 
 	public function komentar_unlock($cat = 1, $id = 0)
 	{
+		if (!$this->web_artikel_model->boleh_ubah($id, $_SESSION['user']))
+			redirect("web/index/$cat");;
+
 		$this->web_artikel_model->komentar_lock($id, 1);
 		redirect("web/index/$cat");
 	}
@@ -217,12 +241,18 @@ class Web extends CI_Controller {
 
 	public function headline($cat = 1, $p = 1, $o = 0, $id = 0)
 	{
+		if (!$this->web_artikel_model->boleh_ubah($id, $_SESSION['user']))
+			redirect("web/index/$cat/$p/$o");;
+
 		$this->web_artikel_model->headline($id);
 		redirect("web/index/$cat/$p/$o");
 	}
 
 	public function slide($cat = 1, $p = 1, $o = 0, $id = 0)
 	{
+		if (!$this->web_artikel_model->boleh_ubah($id, $_SESSION['user']))
+			redirect("web/index/$cat/$p/$o");;
+
 		$this->web_artikel_model->slide($id);
 		redirect("web/index/$cat/$p/$o");
 	}
